backends/azure-functions/requirements.txt
azure-functions
dill